Refactor validation functionality to util file
